+++ packages/ai-react/src/useAiLibState.ts
@@ -1,7 +1,7 @@
 import deepEqual from 'fast-deep-equal/es6/react.js';
 import { useDebugValue, useEffect, useLayoutEffect, useState } from 'react';
 import { useSyncExternalStoreWithSelector } from 'use-sync-external-store/shim/with-selector';
-import { AiLib } from '@connexup/ai-api';
+import { AiLib, EventSourceStatusEnum } from '@connexup/ai-api';
 import { AiLibStateSnapshot, UseAiLibStateOptions } from './types';
 
 const useIsomorphicLayoutEffect =
@@ -10,7 +10,7 @@ const useIsomorphicLayoutEffect =
 export class AiLibStoreManager<TAiLib extends AiLib | null = AiLib | null> {
   private aiLib: TAiLib;
   private lastSnapshot: AiLibStateSnapshot = {
-    status: 'idle',
+    status: EventSourceStatusEnum.IDLE,
     streamMessage: null,
     fullMessages: [],
     error: null,